+++ lib/td_dd/search/structure_enricher.ex
@@ -112,7 +112,11 @@ defmodule TdDd.Search.StructureEnricher do
        when map_size(content) > 0 do
     case Map.get(types, type) do
       %{} = template ->
-        %{structure | search_content: search_content(content, template, domain_ids)}
+        %{
+          structure
+          | search_content: search_content(content, template, domain_ids),
+            alias: Map.get(content, "alias")
+        }
 
       _ ->
         %{structure | search_content: nil}