+++ lib/td_dd_web/views/data_structure_version_view.ex
@@ -8,10 +8,6 @@ defmodule TdDdWeb.DataStructureVersionView do
   alias TdDdWeb.GrantView
   alias TdDqWeb.ImplementationStructureView
 
-  @simplified_note_keys [
-    "alias"
-  ]
-
   def render("show.json", %{actions: actions} = assigns) do
     "show.json"
     |> render(Map.delete(assigns, :actions))
@@ -62,6 +58,7 @@ defmodule TdDdWeb.DataStructureVersionView do
     |> add_grants()
     |> clean_published_note()
     |> Map.take([
+      :alias,
       :ancestry,
       :children,
       :class,
@@ -151,6 +148,7 @@ defmodule TdDdWeb.DataStructureVersionView do
     dsv
     |> add_classes()
     |> Map.take([
+      :alias,
       :data_structure_id,
       :id,
       :name,
@@ -158,12 +156,10 @@ defmodule TdDdWeb.DataStructureVersionView do
       :deleted_at,
       :metadata,
       :classes,
-      :structure_type,
-      :published_note
+      :structure_type
     ])
     |> lift_metadata()
-    |> clean_published_note()
-    |> Map.drop([:structure_type])
+    |> Map.delete(:structure_type)
   end
 
   defp add_children(data_structure_version), do: add_relations(data_structure_version, :children)
@@ -196,12 +192,11 @@ defmodule TdDdWeb.DataStructureVersionView do
   end
 
   defp embedded_relation(%{version: version, relation: relation, relation_type: relation_type}) do
-    structure = data_structure_version_embedded(version)
-
-    Map.new()
-    |> Map.put(:id, relation.id)
-    |> Map.put(:structure, structure)
-    |> Map.put(:relation_type, Map.take(relation_type, [:id, :name, :description]))
+    %{
+      id: relation.id,
+      structure: data_structure_version_embedded(version),
+      relation_type: Map.take(relation_type, [:id, :name, :description])
+    }
   end
 
   defp add_relations(data_structure_version, type) do
@@ -213,13 +208,7 @@ defmodule TdDdWeb.DataStructureVersionView do
         data_structure_version
 
       rs ->
-        relations =
-          Enum.map(rs, fn r ->
-            r
-            |> data_structure_version_embedded()
-            |> simplify_note()
-          end)
-
+        relations = Enum.map(rs, &data_structure_version_embedded/1)
         Map.put(data_structure_version, type, relations)
     end
   end
@@ -240,7 +229,10 @@ defmodule TdDdWeb.DataStructureVersionView do
   defp add_profile(dsv), do: dsv
 
   defp field_structure_json(
-         %{class: "field", data_structure: %{latest_note: latest_note, profile: profile}} = dsv
+         %{
+           class: "field",
+           data_structure: %{latest_note: latest_note, profile: profile}
+         } = dsv
        ) do
     dsv
     |> Map.take([
@@ -253,11 +245,10 @@ defmodule TdDdWeb.DataStructureVersionView do
       :metadata,
       :name,
       :type,
-      :published_note
+      :alias
     ])
     |> lift_metadata()
     |> with_profile_attrs(profile)
-    |> simplify_note()
     |> Map.put(:has_note, not is_nil(latest_note))
   end
 
@@ -296,25 +287,9 @@ defmodule TdDdWeb.DataStructureVersionView do
     Map.put(dsv, :source, source)
   end
 
-  def add_ancestry(%{path: [_ | _] = path} = dsv) do
-    ancestry =
-      path
-      |> Enum.map(&atomize_ancestry/1)
-      |> Enum.map(&clean_published_note/1)
-      |> Enum.map(&simplify_note/1)
-
-    Map.put(dsv, :ancestry, ancestry)
-  end
+  def add_ancestry(%{path: [_ | _] = path} = dsv), do: Map.put(dsv, :ancestry, path)
 
   def add_ancestry(dsv), do: Map.put(dsv, :ancestry, [])
-
-  defp atomize_ancestry(dsv) do
-    published_note = Map.get(dsv, "published_note")
-
-    dsv
-    |> Map.Helpers.atomize_keys()
-    |> Map.put(:published_note, %{df_content: published_note})
-  end
 
   defp lift_metadata(%{metadata: metadata} = dsv) do
     metadata =
@@ -395,34 +370,24 @@ defmodule TdDdWeb.DataStructureVersionView do
 
   defp add_data_structure_type(dsv), do: Map.put(dsv, :data_structure_type, %{})
 
-  defp add_cached_content(dsv, key) do
-    structure = Map.get(dsv, :data_structure)
-
+  defp add_cached_content(%{data_structure: structure} = dsv, key) do
     note =
       structure
       |> Map.get(key, %{})
       |> DataStructures.get_cached_content(dsv)
 
-    structure = Map.put(structure, key, note)
-    Map.put(dsv, :data_structure, structure)
+    %{dsv | data_structure: Map.put(structure, key, note)}
   end
 
   defp clean_published_note(%{published_note: %{df_content: %{} = content}} = dsv) do
-    published_note = DataStructures.get_cached_content(content, dsv)
-    Map.put(dsv, :published_note, published_note)
+    %{dsv | published_note: DataStructures.get_cached_content(content, dsv)}
   end
 
   defp clean_published_note(%{published_note: _} = dsv) do
-    Map.drop(dsv, [:published_note])
+    Map.delete(dsv, :published_note)
   end
 
   defp clean_published_note(dsv), do: dsv
-
-  defp simplify_note(%{published_note: %{} = note} = dsv) do
-    Map.put(dsv, :published_note, Map.take(note, @simplified_note_keys))
-  end
-
-  defp simplify_note(dsv), do: dsv
 
   defp add_tags(ds) do
     tags =